// discord-bot/src/config.d.ts
export declare function isDiscordTokenConfigured(token: string | undefined): boolean;
export declare const config: {
    token: string;
    clientId: string;
    guildId: string | undefined;
    botName: string;
    version: string;
    channels: {
        gameEvents: string;
        moderationLogs: string;
        radioLogs: string;
        chatLogs: string;
    };
};
export default config;
//# sourceMappingURL=config.d.ts.map